% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabulate.R
\name{add_commas}
\alias{add_commas}
\title{Add commas to integers
Add commas to integers (e.g., convert "1000000" to "1,000,000")}
\usage{
add_commas(x, decimals = 0)
}
\arguments{
\item{x}{Integer value}

\item{decimals}{Number of decimal places to print}
}
\value{
Value with commmas.
}
\description{
Add commas to integers
Add commas to integers (e.g., convert "1000000" to "1,000,000")
}
\examples{
\dontrun{
add_commas(x = c(1000, 10000, 100000, 1000000, 10000000, 100000000, 1000000000))
add_commas(x = c(1000, 10000, 100000, 1000000, 10000000, 100000000, 1000000000), decimals = 2)
}
}
\keyword{internal}
