% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{check_wt_type}
\alias{check_wt_type}
\title{Check whether wt_type argument is valid and determine which package to use for weights}
\usage{
check_wt_type(wt_type, generic = FALSE)
}
\arguments{
\item{wt_type}{wt_type argument passed to a meta-analysis function}

\item{generic}{Logical scalar determining whether the effect size is generic (TRUE) or one for which the meta-analysis function estimates error variances (FALSE).}
}
\value{
Character object determining which package should be used to compute weights
}
\description{
Check whether wt_type argument is valid and determine which package to use for weights
}
\keyword{internal}
