% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{clean_warning}
\alias{clean_warning}
\title{Clean warnings and remove warnings present in the environment before running the function of interest}
\usage{
clean_warning(warn_obj1, warn_obj2)
}
\arguments{
\item{warn_obj1}{Initial warning object.}

\item{warn_obj2}{Second warning object.}
}
\value{
Cleaned warning table
}
\description{
Clean warnings and remove warnings present in the environment before running the function of interest
}
\keyword{internal}
