% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{confidence_r}
\alias{confidence_r}
\title{Construct a confidence interval for correlations using Fisher's z transformation}
\usage{
confidence_r(r, n, conf_level = 0.95)
}
\arguments{
\item{r}{A vector of correlations}

\item{n}{A vector of sample sizes}

\item{conf_level}{Confidence level that defines the width of the confidence interval (default = .95).}
}
\value{
A confidence interval of the specified width (or matrix of confidence intervals)
}
\description{
Construct a confidence interval for correlations using Fisher's z transformation
}
\examples{
confidence_r(r = .3, n = 200, conf_level = .95)
}
