% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{convert_consistency2reltype}
\alias{convert_consistency2reltype}
\title{Convert logical variable indicating whether a reliability value is an internal-consistency estimate or a multiple-administration estimate to a string variable of generic reliability types}
\usage{
convert_consistency2reltype(consistency)
}
\arguments{
\item{consistency}{Logical internal-consistency indicators.}
}
\value{
String variable of generic reliability types.
}
\description{
Convert logical variable indicating whether a reliability value is an internal-consistency estimate or a multiple-administration estimate to a string variable of generic reliability types
}
\keyword{internal}
