% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ma.R
\name{convert_pq_to_p}
\alias{convert_pq_to_p}
\title{Convert the variance of a dichotomous variable (i.d., pq) to the proportion of one of the categories in the variable (i.e., p)}
\usage{
convert_pq_to_p(pq)
}
\arguments{
\item{pq}{The variance of a dichotomous variable.}
}
\value{
The proportion of cases in one of the dichotomous groups.
}
\description{
Convert the variance of a dichotomous variable (i.d., pq) to the proportion of one of the categories in the variable (i.e., p)
}
\keyword{internal}
