% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{convert_reltype2consistency}
\alias{convert_reltype2consistency}
\title{Convert string variable containing reliability type indicators to a logical variable indicating whether a reliability value is an internal-consistency estimate or a multiple-administration estimate}
\usage{
convert_reltype2consistency(rel_type, arg_name = NULL)
}
\arguments{
\item{rel_type}{String vector containing reliability type indicators.}

\item{arg_name}{Optional name of the arg_name object.}
}
\value{
Logical internal-consistency indicators.
}
\description{
Convert string variable containing reliability type indicators to a logical variable indicating whether a reliability value is an internal-consistency estimate or a multiple-administration estimate
}
\keyword{internal}
