% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ma.R
\name{convert_sdd_to_sdr}
\alias{convert_sdd_to_sdr}
\title{Convert the SD of d to the SD of r via TSA}
\usage{
convert_sdd_to_sdr(d, sd, p = 0.5)
}
\arguments{
\item{d}{Standardized mean difference in the d-value metric.}

\item{sd}{Standard deviation of the d value.}

\item{p}{Proportion of the dichotomous variable involved in the d value.}
}
\value{
An approximated standard deviation in the correlation metric.
}
\description{
Convert the SD of d to the SD of r via TSA
}
\keyword{internal}
