% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ma.R
\name{convert_sdr_to_sdd}
\alias{convert_sdr_to_sdd}
\title{Convert the SD of r to the SD of d via TSA}
\usage{
convert_sdr_to_sdd(r, sd, p = 0.5)
}
\arguments{
\item{r}{Correlation coefficient.}

\item{sd}{Standard deviation of the correlation.}

\item{p}{Proportion of the dichotomous variable involved in the correlation.}
}
\value{
An approximated standard deviation in the d value metric.
}
\description{
Convert the SD of r to the SD of d via TSA
}
\keyword{internal}
