% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ma.R
\name{convert_varr_to_vard}
\alias{convert_varr_to_vard}
\title{Convert the variance of r to the variance of d via TSA}
\usage{
convert_varr_to_vard(r, var, p)
}
\arguments{
\item{r}{Correlation coefficient.}

\item{var}{Variance of the correlation.}

\item{p}{Proportion of the dichotomous variable involved in the correlation.}
}
\value{
An approximated variance in the d value metric.
}
\description{
Convert the variance of r to the variance of d via TSA
}
\keyword{internal}
