% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_d.R
\name{correct_d_bias}
\alias{correct_d_bias}
\title{Correct for small-sample bias in Cohen's \emph{d} values}
\usage{
correct_d_bias(d, n, a_method = c("gamma", "approx"))
}
\arguments{
\item{d}{Vector of Cohen's d values.}

\item{n}{Vector of sample sizes.}

\item{a_method}{Method used to correct the bias in Cohen's d to convert to Hedges's g. Options are "gamma" (default) for the exact method based on the gamma function (Hedges & Olkin, 1985) or "approx" for the computationally trivial approximation (Borenstein et al., 2006).}
}
\value{
Vector of d values corrected for small-sample bias.
}
\description{
Corrects a vector of Cohen's \emph{d} values for small-sample bias, as Cohen's \emph{d} has a slight positive bias.
}
\details{
The bias correction is estimated as:

\deqn{d_{c}=\frac{d_{obs}}{1+\frac{.75}{n-3}}}{d_c = d / (1 + .75 / (n - 3))}

where \eqn{d_{obs}}{d} is the observed effect size, \eqn{d_{c}}{d_c} is the corrected estimate of \emph{d}, and \emph{n} is the total sample size.
}
\examples{
correct_d_bias(d = .3, n = 30)
correct_d_bias(d = .3, n = 300)
correct_d_bias(d = .3, n = 3000)
}
\references{
Hedges, L. V., & Olkin, I. (1985).
\emph{Statistical methods for meta-analysis}.
Academic Press. p. 104

Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009).
\emph{Introduction to meta-analysis}.
Wiley. p. 27.
}
