% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_r.R
\name{correct_r_uvdrr}
\alias{correct_r_uvdrr}
\title{Correct correlations for Case II direct range restriction and measurement error}
\usage{
correct_r_uvdrr(rxyi, ux = 1, rxx = 1, ryy = 1, ux_observed = TRUE,
  rxx_restricted = TRUE, ryy_restricted = TRUE, n = NULL,
  conf_level = 0.95, correct_bias = FALSE)
}
\arguments{
\item{rxyi}{Vector of observed correlations.}

\item{ux}{Vector of u ratios for X.}

\item{rxx}{Vector of reliability coefficients for X.}

\item{ryy}{Vector of reliability coefficients for Y.}

\item{ux_observed}{Logical vector in which each entry specifies whether the corresponding ux value is an observed-score u ratio (\code{TRUE}) or a true-score u ratio. All entries are \code{TRUE} by default.}

\item{rxx_restricted}{Logical vector in which each entry specifies whether the corresponding rxx value is an incumbent reliability (\code{TRUE}) or an applicant reliability. All entries are \code{TRUE} by default.}

\item{ryy_restricted}{Logical vector in which each entry specifies whether the corresponding rxx value is an incumbent reliability (\code{TRUE}) or an applicant reliability. All entries are \code{TRUE} by default.}

\item{n}{Optional vector of sample sizes associated with the rxyi correlations.}

\item{conf_level}{Confidence level to define the width of the confidence interval (default = .95).}

\item{correct_bias}{Logical argument that determines whether to correct error-variance estimates for small-sample bias in correlations (\code{TRUE}) or not (\code{FALSE}).
For sporadic corrections (e.g., in mixed artifact-distribution meta-analyses), this should be set to \code{FALSE}, the default).}
}
\value{
Data frame(s) of observed correlations (\code{rxyi}), operational range-restricted correlations corrected for measurement error in Y only (\code{rxpi}), operational range-restricted correlations corrected for measurement error in X only (\code{rtyi}), and range-restricted true-score correlations (\code{rtpi}),
range-corrected observed-score correlations (\code{rxya}), operational range-corrected correlations corrected for measurement error in Y only (\code{rxpa}), operational range-corrected correlations corrected for measurement error in X only (\code{rtya}), and range-corrected true-score correlations (\code{rtpa}).
}
\description{
Correct correlations for Case II direct range restriction and measurement error
}
\details{
\deqn{\left[\frac{\rho_{XY_{i}}}{u_{X}\sqrt{\rho_{YY_{i}}}\sqrt{\left(\frac{1}{u_{X}^{2}}-1\right)\frac{\rho_{XY_{i}}^{2}}{\rho_{YY_{i}}}+1}}\right]/\sqrt{\rho_{XX_{a}}}}else{(rxyi / (ux * sqrt(ryyi) * sqrt((1 / ux^2 - 1) * rxyi^2 / ryyi + 1))) / sqrt(rxxa)}
}
\examples{
correct_r_uvdrr(rxyi = .3, ux = .8, rxx = .8, ryy = .8,
     ux_observed = TRUE, rxx_restricted = TRUE, ryy_restricted = TRUE)
correct_r_uvdrr(rxyi = .3, ux = .8, rxx = .8, ryy = .8,
     ux_observed = TRUE, rxx_restricted = TRUE, ryy_restricted = TRUE, n = 100)
}
\references{
Hunter, J. E., Schmidt, F. L., & Le, H. (2006).
Implications of direct and indirect range restriction for meta-analysis methods and findings.
\emph{Journal of Applied Psychology, 91}(3), 594–612. \url{https://doi.org/10.1037/0021-9010.91.3.594}
}
\keyword{internal}
