% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_d_bb_multi}
\alias{data_d_bb_multi}
\title{Hypothetical \emph{d} value dataset simulated with sampling error only}
\format{
data.frame
}
\usage{
data(data_d_bb_multi)
}
\description{
Data set for use in example meta-analyses of multiple variables.
}
\examples{
data(data_d_bb_multi)
}
\keyword{datasets}
