% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_r_bvirr}
\alias{data_r_bvirr}
\title{Hypothetical dataset simulated to statisfy the assumptions of the bivariate correction for indirect range restriction}
\format{data.frame}
\usage{
data(data_r_bvirr)
}
\description{
Data set for use in example meta-analyses of bivariate indirect range restriction.
}
\examples{
data(data_r_bvirr)
}
\keyword{datasets}
