% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_r_mcdaniel_1994}
\alias{data_r_mcdaniel_1994}
\title{Artifact-distribution meta-analysis of the validity of interviews}
\format{
data.frame
}
\usage{
data(data_r_mcdaniel_1994)
}
\description{
Data set to demonstrate corrections for univariate range restriction and criterion measurement error using artifact distributions.
}
\examples{
data(data_r_mcdaniel_1994)
}
\references{
McDaniel, M. A., Whetzel, D. L., Schmidt, F. L., & Maurer, S. D. (1994).
The validity of employment interviews: A comprehensive review and meta-analysis.
\emph{Journal of Applied Psychology, 79}(4), 599–616. \doi{10.1037/0021-9010.79.4.599}
}
\keyword{datasets}
