% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_r_mcleod_2007}
\alias{data_r_mcleod_2007}
\title{Bare-bones meta-analysis of parenting and childhood depression}
\format{
data.frame
}
\usage{
data(data_r_mcleod_2007)
}
\description{
Data set to demonstrate bare-bones meta-analysis.
}
\examples{
data(data_r_mcleod_2007)
}
\references{
McLeod, B. D., Weisz, J. R., & Wood, J. J., (2007).
Examining the association between parenting and childhood depression:  A meta-analysis.
\emph{Clinical Psychology Review, 27}(8), 986–1003. \doi{10.1016/j.cpr.2007.03.001}
}
\keyword{datasets}
