% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_r_meas}
\alias{data_r_meas}
\title{Hypothetical dataset simulated to satisfy the assumptions of the correction for measurement error only}
\format{
data.frame
}
\usage{
data(data_r_meas)
}
\description{
Data set for use in example meta-analyses correcting for measurement error in two variables.
}
\examples{
data(data_r_meas)
}
\keyword{datasets}
