% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_r_meas_multi}
\alias{data_r_meas_multi}
\title{Hypothetical correlation dataset simulated to satisfy the assumptions of the correction for measurement error only in multiple constructs}
\format{data.frame}
\usage{
data(data_r_meas_multi)
}
\description{
Data set for use in example meta-analyses correcting for measurement error in multiple variables.
}
\examples{
data(data_r_meas_multi)
}
\keyword{datasets}
