% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_r_oh_2009}
\alias{data_r_oh_2009}
\title{Second order meta-analysis of operational validities of big five personality measures across East Asian countries}
\format{
data.frame
}
\usage{
data(data_r_oh_2009)
}
\description{
Example of a second-order meta-analysis of correlations corrected using artifact-distribution methods.
}
\examples{
data(data_r_oh_2009)
}
\references{
Oh, I. -S. (2009).
\emph{The Five-Factor Model of personality and job performance in East Asia: A cross-cultural validity generalization study}.
(Doctoral dissertation) Iowa City, IA: University of Iowa. \url{https://search.proquest.com/dissertations/docview/304903943/}

Schmidt, F. L., & Oh, I.-S. (2013).
Methods for second order meta-analysis and illustrative applications.
\emph{Organizational Behavior and Human Decision Processes, 121}(2), 204–218. \url{https://doi.org/10.1016/j.obhdp.2013.03.002}
}
\keyword{datasets}
