% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_r_roth_2015}
\alias{data_r_roth_2015}
\title{Artifact-distribution meta-analysis of the correlation between school grades and cognitive ability}
\format{
data.frame
}
\usage{
data(data_r_roth_2015)
}
\description{
Data set to demonstrate corrections for univariate range restriction and cognitive ability measurement error.
}
\examples{
data(data_r_roth_2015)
}
\references{
Roth, B., Becker, N., Romeyke, S., Schäfer, S., Domnick, F., & Spinath, F. M. (2015).
Intelligence and school grades: A meta-analysis.
\emph{Intelligence, 53}, 118–137. \url{https://doi.org/10.1016/j.intell.2015.09.002}
}
\keyword{datasets}
