% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_r_uvdrr}
\alias{data_r_uvdrr}
\title{Hypothetical dataset simulated to satisfy the assumptions of the univariate correction for direct range restriction}
\format{data.frame}
\usage{
data(data_r_uvdrr)
}
\description{
Data set for use in example meta-analyses correcting for univariate direct range restriction.
}
\examples{
data(data_r_uvdrr)
}
\keyword{datasets}
