% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_r_uvirr}
\alias{data_r_uvirr}
\title{Hypothetical dataset simulated to satisfy the assumptions of the univariate correction for indirect range restriction}
\format{
data.frame
}
\usage{
data(data_r_uvirr)
}
\description{
Data set for use in example meta-analyses correcting for univariate indirect range restriction.
}
\examples{
data(data_r_uvirr)
}
\keyword{datasets}
