% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_r_internal.R
\name{.attenuate_r_bvdrr}
\alias{.attenuate_r_bvdrr}
\title{Internal function to undo the Case V correction for bivariate indirect range restriction}
\usage{
.attenuate_r_bvdrr(rtpa, qxa = 1, qya = 1, ux = 1, uy = 1)
}
\arguments{
\item{rtpa}{Correlation corrected for range restriction and measurement error.}

\item{qxa}{Vector of square-root of applicant reliability coefficients for X.}

\item{qya}{Vector of square-root of applicant reliability coefficients for Y.}

\item{ux}{Vector of observed-score u ratios for X.}

\item{uy}{Vector of observed-score u ratios for Y.}
}
\value{
A vector of attenuated correlations.
}
\description{
Internal function to undo the Case V correction for bivariate indirect range restriction
}
\keyword{internal}
