% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_r_internal.R
\name{.attenuate_r_rb}
\alias{.attenuate_r_rb}
\title{Internal function to attenuate correlations using Raju and Burke's formula for univariate direct range restriction}
\usage{
.attenuate_r_rb(rtpa, qx = 1, qy = 1, ux = 1)
}
\arguments{
\item{rtpa}{Vector of fully corrected correlations.}

\item{qx}{Vector of square-root of reliability coefficients for X.}

\item{qy}{Vector square-root of reliability coefficients for Y.}

\item{ux}{Vector of observed-score u ratios for X.}
}
\value{
A vector of corrected correlations.
}
\description{
Internal function to attenuate correlations using Raju and Burke's formula for univariate direct range restriction
}
\keyword{internal}
