% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_r_internal.R
\name{.attenuate_r_uvirr}
\alias{.attenuate_r_uvirr}
\title{Internal function to undo the Case IV correction for univariate indirect range restriction}
\usage{
.attenuate_r_uvirr(rtpa, qxi = 1, qyi = 1, ut = 1)
}
\arguments{
\item{rtpa}{Vector of fully corrected correlations.}

\item{qxi}{Vector of square-root of incumbent reliability coefficient for X.}

\item{qyi}{Vector of square-root of incumbent reliability coefficient for Y.}

\item{ut}{Vector of true-score u ratios for X.}
}
\value{
A vector of attenuated correlations.
}
\description{
Internal function to undo the Case IV correction for univariate indirect range restriction
}
\keyword{internal}
