% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_r_internal.R
\name{.correct_r_bvirr}
\alias{.correct_r_bvirr}
\title{Internal function to compute the Case V correction for bivariate indirect range restriction}
\usage{
.correct_r_bvirr(
  rxyi,
  qxa = 1,
  qya = 1,
  ux = 1,
  uy = 1,
  sign_rxz = 1,
  sign_ryz = 1
)
}
\arguments{
\item{rxyi}{Vector of observed correlations.}

\item{qxa}{Vector of square-root of applicant reliability coefficients for X.}

\item{qya}{Vector of square-root of applicant reliability coefficients for Y.}

\item{ux}{Vector of observed-score u ratios for X.}

\item{uy}{Vector of observed-score u ratios for Y.}

\item{sign_rxz}{Sign of the relationship between X and the selection mechanism.}

\item{sign_ryz}{Sign of the relationship between Y and the selection mechanism.}
}
\value{
A vector of corrected correlations.
}
\description{
Internal function to compute the Case V correction for bivariate indirect range restriction
}
\keyword{internal}
