% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_r_internal.R
\name{.correct_r_rb}
\alias{.correct_r_rb}
\title{Internal function to compute Raju and Burke's correction for univariate direct range restriction}
\usage{
.correct_r_rb(rxyi, qx = 1, qy = 1, ux = 1)
}
\arguments{
\item{rxyi}{Vector of observed correlations.}

\item{qx}{Vector of square-root of reliability coefficients for X.}

\item{qy}{Vector square-root of reliability coefficients for Y.}

\item{ux}{Vector of observed-score u ratios for X.}
}
\value{
A vector of corrected correlations.
}
\description{
Internal function to compute Raju and Burke's correction for univariate direct range restriction
}
\keyword{internal}
