% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_r_internal.R
\name{.correct_r_uvdrr}
\alias{.correct_r_uvdrr}
\title{Internal function to compute the Case II correction for univariate direct range restriction}
\usage{
.correct_r_uvdrr(rxyi, qxa = 1, qyi = 1, ux = 1)
}
\arguments{
\item{rxyi}{Vector of observed correlations.}

\item{qxa}{Vector of square-root of applicant reliability coefficients for X.}

\item{qyi}{Vector square-root of incumbent reliability coefficients for Y.}

\item{ux}{Vector of observed-score u ratios for X.}
}
\value{
A vector of corrected correlations.
}
\description{
Internal function to compute the Case II correction for univariate direct range restriction
}
\keyword{internal}
