% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_r_internal.R
\name{.correct_r_uvirr}
\alias{.correct_r_uvirr}
\title{Internal function to compute the Case IV correction for univariate indirect range restriction}
\usage{
.correct_r_uvirr(rxyi, qxi = 1, qyi = 1, ut = 1)
}
\arguments{
\item{rxyi}{Vector of observed correlations.}

\item{qxi}{Vector of square-root of incumbent reliability coefficient for X.}

\item{qyi}{Vector of square-root of incumbent reliability coefficient for Y.}

\item{ut}{Vector of true-score u ratio for X.}
}
\value{
A vector of corrected correlations.
}
\description{
Internal function to compute the Case IV correction for univariate indirect range restriction
}
\keyword{internal}
