% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_r.R
\name{.descriptives_database}
\alias{.descriptives_database}
\title{Compute weighted descriptive statistics for a database}
\usage{
.descriptives_database(dat, wt)
}
\arguments{
\item{dat}{Numeric matrix or data frame}

\item{wt}{Vector of weights to be applied to all columns of dat}
}
\value{
A matrix of weighted descriptive statistics
}
\description{
Compute weighted descriptive statistics for a database
}
\keyword{internal}
