% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{.heterogeneity}
\alias{.heterogeneity}
\title{Computation of heterogeneity indices from meta-analytic results}
\usage{
.heterogeneity(mean_es, var_es, var_e, var_art = NA, var_pre = NA,
  wt_vec, N, k, es_vec, es_failsafe = NULL, conf_level = 0.95,
  es_type = "es")
}
\arguments{
\item{mean_es}{The mean effect size.}

\item{var_es}{The observed variances of effect sizes.}

\item{var_e}{The predicted error variance in effect sizes.}

\item{var_art}{The variance of effect sizes predicted from artifacts.}

\item{var_pre}{The total amount of artifactual variance predicted from artifacts and statistical error,}

\item{wt_vec}{The vector of weights used in the meta-analysis.}

\item{N}{The total sample size of the meta-analysis.}

\item{k}{The number of effect sizes included in the meta-analysis.}

\item{es_vec}{The vector of effect sizes used in the meta-analysis.}

\item{es_failsafe}{Failsafe value of the effect size for use in file-drawer analyses.}

\item{conf_level}{Confidence level to define the width of the confidence interval (default = .95).}

\item{es_type}{Name of effect-size type.}
}
\value{
A list of heterogeneity statistics.
}
\description{
Computation of heterogeneity indices from meta-analytic results
}
\keyword{internal}
