% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ma.R
\name{.identify_ma_cols}
\alias{.identify_ma_cols}
\title{Identify meta-analysis type and provide new column names for a meta-analysis}
\usage{
.identify_ma_cols(col_names)
}
\arguments{
\item{col_names}{Column names of a meta-analysis table.}
}
\value{
Meta-analysis type, old column names of table, column names of table after effect-size conversion, and a vector categorizing the types of entries supplied in the table.
}
\description{
Identify meta-analysis type and provide new column names for a meta-analysis
}
\keyword{internal}
