% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ad_list.R
\name{.ma_artifacts}
\alias{.ma_artifacts}
\title{Internal function for computing meta-analyses of artifacts}
\usage{
.ma_artifacts(data, ma_arg_list)
}
\arguments{
\item{data}{Data frame of bare-bones information.}

\item{ma_arg_list}{List of arguments to be used in the meta-analysis function.}
}
\value{
A list object containing the results of bare-bones meta-analyses of correlations.
}
\description{
Internal function for computing meta-analyses of artifacts
}
\keyword{internal}
