% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{.ma_bootstrap}
\alias{.ma_bootstrap}
\title{Wrapper function to facilitate bootstrapped meta-analyses}
\usage{
.ma_bootstrap(
  data,
  ma_fun_boot,
  boot_iter = 1000,
  boot_conf_level = 0.95,
  boot_ci_type = "norm",
  ma_arg_list,
  convert_ma = FALSE
)
}
\arguments{
\item{data}{Data to be meta-analyzed.}

\item{ma_fun_boot}{Meta-analysis function.}

\item{boot_iter}{Number of bootstrap iterations to be computed.}

\item{boot_conf_level}{Width of confidence intervals to be constructed for all bootstrapped statistics.}

\item{boot_ci_type}{Type of bootstrapped confidence interval (see "type" options for boot::boot.ci for possible arguments).}

\item{ma_arg_list}{List of arguments to be passed to the meta-analysis function.}
}
\value{
A list containing (1) a summary matrix of means, variances, and confidence intervals of bootstrapped values and (2) the raw
output of the bootstrapping function.
}
\description{
Wrapper function to facilitate bootstrapped meta-analyses
}
\keyword{internal}
