% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{.ma_cumulative}
\alias{.ma_cumulative}
\title{Cumulative meta-analyses}
\usage{
.ma_cumulative(
  data,
  sort_method = c("n", "inv_var", "weight"),
  ma_fun_boot,
  ma_arg_list
)
}
\arguments{
\item{data}{Data to be meta-analyzed.}

\item{sort_method}{Method to sort samples in the cumulative meta-analysis. Options are "weight" to sort by weight (default), "n" to sort by sample size, and "inv_var" to sort by inverse variance.}

\item{ma_fun_boot}{Meta-analysis function.}

\item{ma_arg_list}{List of arguments to be passed to the meta-analysis function.}
}
\value{
Cumulative meta-analysis table
}
\description{
Cumulative meta-analyses
}
\examples{
## Analysis TBD
}
\keyword{internal}
