% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_d_bb.R
\name{.ma_d_bb}
\alias{.ma_d_bb}
\title{Internal function for computing bare-bones meta-analyses of d values}
\usage{
.ma_d_bb(data, ma_arg_list, run_lean = FALSE)
}
\arguments{
\item{data}{Data frame of bare-bones information.}

\item{ma_arg_list}{List of arguments to be used in the meta-analysis function.}

\item{run_lean}{If TRUE, the meta-analysis will not generate an escalc object. Meant to speed up bootstrap analyses that do not require supplmental output.}
}
\value{
A list object containing the results of bare-bones meta-analyses of d values.
}
\description{
Internal function for computing bare-bones meta-analyses of d values
}
\keyword{internal}
