% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_generic.R
\name{.ma_generic}
\alias{.ma_generic}
\title{Internal function for computing bare-bones meta-analyses of generic effect sizes}
\usage{
.ma_generic(data, run_lean = FALSE, ma_arg_list)
}
\arguments{
\item{data}{Data frame of bare-bones information.}

\item{run_lean}{If TRUE, the meta-analysis will not generate an escalc object. Meant to speed up bootstrap analyses that do not require supplemental output.}

\item{ma_arg_list}{List of arguments to be used in the meta-analysis function.}
}
\value{
A list object containing the results of bare-bones meta-analyses of generic effect sizes.
}
\description{
Internal function for computing bare-bones meta-analyses of generic effect sizes
}
\keyword{internal}
