% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_generic.R
\name{.ma_generic_boot}
\alias{.ma_generic_boot}
\title{Internal function for computing bootstrapped bare-bones meta-analyses of generic effect sizes}
\usage{
.ma_generic_boot(data, i, ma_arg_list)
}
\arguments{
\item{data}{Data frame of bare-bones information.}

\item{i}{Vector of indexes to select studies from 'data'.}

\item{ma_arg_list}{List of arguments to be passed to the meta-analysis function.}
}
\value{
A list object containing the results of bootstrapped bare-bones meta-analyses of generic effect sizes.
}
\description{
Internal function for computing bootstrapped bare-bones meta-analyses of generic effect sizes
}
\keyword{internal}
