% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{.ma_leave1out}
\alias{.ma_leave1out}
\title{Leave-one-out (i.e., jackknife) meta-analyses}
\usage{
.ma_leave1out(data, ma_fun_boot, ma_arg_list)
}
\arguments{
\item{data}{Data to be meta-analyzed.}

\item{ma_fun_boot}{Meta-analysis function.}

\item{ma_arg_list}{List of arguments to be passed to the meta-analysis function.}
}
\value{
Leave-one-out results for the specified meta-analysis
}
\description{
Leave-one-out (i.e., jackknife) meta-analyses
}
\examples{
## Analysis TBD
}
\keyword{internal}
