% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_r_ic.R
\name{.ma_r_ic}
\alias{.ma_r_ic}
\title{Internal function for computing individual-correction meta-analyses of correlations}
\usage{
.ma_r_ic(data, type = "all", run_lean = FALSE, ma_arg_list)
}
\arguments{
\item{data}{Data frame of individual-correction information.}

\item{type}{Type of correlation to be meta-analyzed: "ts" for true score, "vgx" for validity generalization with "X" as the predictor,
"vgy" for for validity generalization with "X" as the predictor, and "all" for the complete set of results.}

\item{run_lean}{If TRUE, the meta-analysis will not generate an escalc object. Meant to speed up bootstrap analyses that do not require supplemental output.}

\item{ma_arg_list}{List of arguments to be used in the meta-analysis function.}
}
\value{
A list object containing the results of individual-correction meta-analyses of correlations.
}
\description{
Internal function for computing individual-correction meta-analyses of correlations
}
\keyword{internal}
