% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_r_order2.R
\name{.ma_r_order2_ic_boot}
\alias{.ma_r_order2_ic_boot}
\title{Internal function for computing bootstrapped second-order individual-correction meta-analyses of correlations}
\usage{
.ma_r_order2_ic_boot(data, i, ma_arg_list)
}
\arguments{
\item{data}{Data frame of meta-analytic information.}

\item{i}{Vector of indexes to select studies from 'data'.}

\item{ma_arg_list}{List of arguments to be passed to the meta-analysis function.}
}
\value{
A list object containing the results of bootstrapped second-order individual-correction meta-analyses of correlations
}
\description{
Internal function for computing bootstrapped second-order individual-correction meta-analyses of correlations
}
\keyword{internal}
