% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{.metabulate}
\alias{.metabulate}
\title{Internal function for .metabulating results tables}
\usage{
.metabulate(meta_tables, ma_type = "ad_ts", output_format = "word",
  caption = caption, show_msd = TRUE, show_conf = TRUE,
  show_cred = TRUE, show_se = FALSE, show_var = FALSE,
  collapse_construct_labels = TRUE, bold_headers = TRUE,
  es_type = NULL, symbol_es = "ES", digits = 2L,
  decimal.mark = getOption("OutDec"), leading0 = "figure",
  neg.sign = "&minus;", pos.sign = "figure", drop0integer = TRUE,
  big.mark = "&#8239;", big.interval = 3L, small.mark = "&#8239;",
  small.interval = 3L, na.mark = "&mdash;", lgl.mark = c("+",
  "&minus;"), inf.mark = c("&#8199;&infin;", "&minus;&infin;"),
  conf_format = "parentheses", cred_format = "parentheses",
  verbose = FALSE, unicode = unicode, conf_level = 0.95,
  cred_level = 0.8)
}
\description{
Internal function for .metabulating results tables
}
\keyword{internal}
