% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{.plot_forest_meta}
\alias{.plot_forest_meta}
\title{Internal plotting function for forest plots}
\usage{
.plot_forest_meta(ma_mat, ma_vec = NULL, analysis = "leave1out")
}
\arguments{
\item{ma_mat}{Matrix of meta-analytic results to be plotted.}

\item{ma_vec}{An optional vector of overall meta-analytic results to use as reference points on the plots.}

\item{analysis}{Type of analysis to be plotted: leave-one-out or cumulative.}
}
\value{
A list of forest plots
}
\description{
Internal plotting function for forest plots
}
\keyword{internal}
