% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.plot_funnel}
\alias{.plot_funnel}
\title{Internal funnel-plot generator}
\usage{
.plot_funnel(x)
}
\arguments{
\item{x}{An escalc-class object}
}
\value{
A funnel plot.
}
\description{
Internal funnel-plot generator
}
\author{
John Sakaluk
}
\keyword{internal}
