% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_print.R
\name{.print.summary.lm_mat}
\alias{.print.summary.lm_mat}
\title{Print method for objects of the class "summary.lm_mat"}
\usage{
.print.summary.lm_mat(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\description{
Print method for objects of the class "summary.lm_mat"
}
\keyword{internal}
