% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_r.R
\name{.rbeta}
\alias{.rbeta}
\title{Generate values from a beta distribution, given a mean and standard deviation of the distribution}
\usage{
.rbeta(n, mean, sd)
}
\arguments{
\item{n}{Number of values to generate}

\item{mean}{Mean of the distribution}

\item{sd}{Standard deviation of the distribution}
}
\value{
A vector of simulated values from a distribution bounded at 0 and 1
}
\description{
Generate values from a beta distribution, given a mean and standard deviation of the distribution
}
\examples{
\dontrun{
.rbeta(n = 10, mean = .8, sd, .2)
}
}
\keyword{internal}
