% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{.tau_squared_m_solver}
\alias{.tau_squared_m_solver}
\title{tau_m_squared Solver}
\usage{
.tau_squared_m_solver(Q_m, wi, k)
}
\arguments{
\item{Q_m}{The Q_r statistic.}

\item{wi}{Vector of inverse within-study sampling variances.}

\item{k}{The number of effect sizes included in the meta-analysis.}
}
\value{
tau_r_squared
}
\description{
Function to solve for tau_m_squared (outlier-robust estimator of tau_squared based on absolute deviations from median)
}
\author{
Lifeng Lin and Haitao Chu
}
