% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{.tau_squared_r_solver}
\alias{.tau_squared_r_solver}
\title{tau_r_squared Solver}
\usage{
.tau_squared_r_solver(Q_r, wi)
}
\arguments{
\item{Q_r}{The Q_r statistic.}

\item{wi}{Vector of inverse within-study sampling variances.}
}
\value{
tau_r_squared
}
\description{
Function to solve for tau_r_squared (outlier-robust estimator of tau_squared based on absolute deviations from mean)
}
\author{
Lifeng Lin achind Haitao Chu
}
\keyword{internal}
