% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dplyr_ma_psychmeta.R
\name{filter}
\alias{filter}
\title{Return rows with matching conditions}
\usage{
filter(.data, ...)
}
\arguments{
\item{.data}{A tbl. All main verbs are S3 generics and provide methods for tbl_df(), dtplyr::tbl_dt() and dbplyr::tbl_dbi().}

\item{...}{Logical predicates defined in terms of the variables in .data. Multiple conditions are combined with &. Only rows where the condition evaluates to TRUE are kept.}
}
\description{
Use filter() find rows/cases where conditions are true. Unlike base subsetting with [, rows where the condition evaluates to NA are dropped.
See the dplyr::filter help for more information about the filter() function.
}
\keyword{internal}
