% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_listnonnull}
\alias{filter_listnonnull}
\title{Filter a list to remove NULL entries}
\usage{
filter_listnonnull(x)
}
\arguments{
\item{x}{A list}
}
\value{
A list with NULL entries removed.
}
\description{
Filter a list to remove NULL entries
}
\keyword{internal}
