% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_r}
\alias{filter_r}
\title{Filter to detect and remove impossible values in vectors of correlations and sample sizes.}
\usage{
filter_r(r_vec, n_vec)
}
\arguments{
\item{r_vec}{Vector of correlations.}

\item{n_vec}{Vector of sample sizes.}
}
\value{
List of filtered correlations and sample sizes.
}
\description{
Filter to detect and remove impossible values in vectors of correlations and sample sizes.
}
\examples{
## Not run
## filter_r(r_vec = c(-.3, .5, 1.1), n_vec = c(100, 100, 100))
## filter_r(r_vec = c(-.3, .5, .8), n_vec = c(Inf, 100, 100))
## filter_r(r_vec = c(-.3, .5, .8), n_vec = c(2, 100, 100))
}
\keyword{internal}
