% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_r_bar}
\alias{filter_r_bar}
\title{Filter to detect and remove impossible values in vectors of meta-analytic mean correlations and numbers of studies}
\usage{
filter_r_bar(r_bar_vec, k_vec)
}
\arguments{
\item{r_bar_vec}{Vector of meta-analytic mean correlations.}

\item{k_vec}{Vector of numbers of studies (k values).}
}
\value{
List of filtered correlations and numbers of studies.
}
\description{
Filter to detect and remove impossible values in vectors of meta-analytic mean correlations and numbers of studies
}
\examples{
## Not run
## filter_r_bar(r_bar_vec = c(-.3, .5, 1.1), k_vec = c(100, 100, 100))
## filter_r_bar(r_bar_vec = c(-.3, .5, .8), k_vec = c(Inf, 100, 100))
## filter_r_bar(r_bar_vec = c(-.3, .5, .8), k_vec = c(2, 100, 100))
}
\keyword{internal}
