% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_rel}
\alias{filter_rel}
\title{Filter to remove impossible values from vectors of reliabilities and corresponding weights.}
\usage{
filter_rel(rel_vec, wt_vec)
}
\arguments{
\item{rel_vec}{Vector of reliability estimates.}

\item{wt_vec}{Vector of weights corresponding to the elements of rel_vec.}
}
\value{
List of filtered reliabilities and weights.
}
\description{
Filter to remove impossible values from vectors of reliabilities and corresponding weights.
}
\examples{
## Not run
## filter_rel(rel_vec = c(0, .8), wt_vec = c(80, 100))
## filter_rel(rel_vec = c(.7, .8), wt_vec = c(-80, 100))
}
\keyword{internal}
