% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_u}
\alias{filter_u}
\title{Filter to remove impossible values from vectors of u ratios and corresponding weights.}
\usage{
filter_u(u_vec, wt_vec)
}
\arguments{
\item{u_vec}{Vector of u ratios.}

\item{wt_vec}{Vector of weights corresponding to the elements of u_vec}
}
\value{
List of filtered u ratios and weights.
}
\description{
Filter to remove impossible values from vectors of u ratios and corresponding weights.
}
\examples{
## Not run
## filter_u(u_vec = c(0, .8), wt_vec = c(80, 100))
## filter_u(u_vec = c(.7, .8), wt_vec = c(-80, 100))
}
\keyword{internal}
