% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_directory}
\alias{generate_directory}
\title{Generate a system of folders from a file path to a new directory}
\usage{
generate_directory(path)
}
\arguments{
\item{path}{The path to the directory to be created}
}
\value{
Creates a system of folders to a new directory
}
\description{
This function is intended to be helpful in simulations when directories need to be created and named according to values generated during the simulation.
}
\keyword{utilities}
