% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{interval_warning}
\alias{interval_warning}
\title{Warning message for the widths of uncertainty intervals}
\usage{
interval_warning(interval, interval_name = NULL, default)
}
\arguments{
\item{interval}{Width for interval}

\item{interval_name}{Name of interval}

\item{default}{Default value for interval}
}
\value{
Warning if length of 'interval' takes on impossible values and the revised value of 'interval'.
}
\description{
Warning message for the widths of uncertainty intervals
}
\keyword{internal}
