% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{limits_tau}
\alias{limits_tau}
\title{Confidence limits of tau}
\usage{
limits_tau(Q, df, C, conf_level = 0.95)
}
\arguments{
\item{Q}{The Q statistic from the meta-analysis.}

\item{df}{The degrees of freedom associated with the Q statistic.}

\item{C}{The statistic computed as: sum(weights) - (sum(weights^2) / sum(weights))}

\item{conf_level}{Confidence level}
}
\value{
The confidence limits of tau
}
\description{
Confidence limits of tau
}
\keyword{internal}
