% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_r_ad_meas.R
\name{ma_r_ad.int_meas}
\alias{ma_r_ad.int_meas}
\title{Interactive artifact-distribution meta-analysis correcting for measurement error}
\usage{
ma_r_ad.int_meas(x)
}
\arguments{
\item{x}{List of bare-bones meta-analytic data, artifact-distribution objects for X and Y, and other meta-analysis options.}
}
\value{
A meta-analysis class object containing all results.
}
\description{
Interactive artifact-distribution meta-analysis correcting for measurement error
}
\references{
Schmidt, F. L., & Hunter, J. E. (2015).
\emph{Methods of meta-analysis: Correcting error and bias in research findings (3rd ed.)}.
Thousand Oaks, California: SAGE Publications, Inc. Chapter 4.
}
\keyword{internal}
