% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_r_ad_none.R
\name{ma_r_ad.int_none}
\alias{ma_r_ad.int_none}
\alias{ma_r_ad.tsa_none}
\title{Null artifact distribution result: No corrections performed}
\usage{
ma_r_ad.int_none(x)

ma_r_ad.tsa_none(x)
}
\arguments{
\item{x}{List of bare-bones meta-analytic data, artifact-distribution objects for X and Y, and other meta-analysis options.}
}
\value{
A meta-analysis class object containing all results.
}
\description{
Null artifact distribution result: No corrections performed
}
\keyword{internal}
