% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_r_ad_rb_adj.R
\name{ma_r_ad.int_rbAdj}
\alias{ma_r_ad.int_rbAdj}
\title{Interactive artifact-distribution meta-analysis correcting for Case II direct range restriction and measurement error}
\usage{
ma_r_ad.int_rbAdj(x)
}
\arguments{
\item{x}{List of bare-bones meta-analytic data, artifact-distribution objects for X and Y, and other meta-analysis options.}
}
\value{
A meta-analysis class object containing all results.
}
\description{
Interactive artifact-distribution meta-analysis correcting for Case II direct range restriction and measurement error
}
\references{
Schmidt, F. L., & Hunter, J. E. (2015).
\emph{Methods of meta-analysis: Correcting error and bias in research findings} (3rd ed.).
Thousand Oaks, CA: Sage. \url{https://doi.org/10/b6mg}. Chapter 4.

Law, K. S., Schmidt, F. L., & Hunter, J. E. (1994).
Nonlinearity of range corrections in meta-analysis: Test of an improved procedure.
\emph{Journal of Applied Psychology, 79}(3), 425–438. \url{https://doi.org/10.1037/0021-9010.79.3.425}

Raju, N. S., & Burke, M. J. (1983).
Two new procedures for studying validity generalization.
\emph{Journal of Applied Psychology, 68}(3), 382–395. \url{https://doi.org/10.1037/0021-9010.68.3.382}
}
\keyword{internal}
