% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_r_ad_uvirr.R
\name{ma_r_ad.int_uvirr}
\alias{ma_r_ad.int_uvirr}
\title{Interactive artifact-distribution meta-analysis correcting for Case IV indirect range restriction and measurement error}
\usage{
ma_r_ad.int_uvirr(x)
}
\arguments{
\item{x}{List of bare-bones meta-analytic data, artifact-distribution objects for X and Y, and other meta-analysis options.}
}
\value{
A meta-analysis class object containing all results.
}
\description{
Interactive artifact-distribution meta-analysis correcting for Case IV indirect range restriction and measurement error
}
\references{
Le, H., & Schmidt, F. L. (2006).
Correcting for indirect range restriction in meta-analysis: Testing a new meta-analytic procedure.
\emph{Psychological Methods, 11}(4), 416–438. \doi{10.1037/1082-989X.11.4.416}

Schmidt, F. L., & Hunter, J. E. (2015).
\emph{Methods of meta-analysis: Correcting error and bias in research findings} (3rd ed.).
Thousand Oaks, CA: Sage. \doi{10/b6mg}. Chapter 4.

Law, K. S., Schmidt, F. L., & Hunter, J. E. (1994).
Nonlinearity of range corrections in meta-analysis: Test of an improved procedure.
\emph{Journal of Applied Psychology, 79}(3), 425–438. \doi{10.1037/0021-9010.79.3.425}
}
\keyword{internal}
