% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_r_ad_bvirr.R
\name{ma_r_ad.tsa_bvirr}
\alias{ma_r_ad.tsa_bvirr}
\title{Taylor series approximation artifact-distribution meta-analysis correcting for Case V indirect range restriction and measurement error}
\usage{
ma_r_ad.tsa_bvirr(x)
}
\arguments{
\item{x}{List of bare-bones meta-analytic data, artifact-distribution objects for X and Y, and other meta-analysis options.}
}
\value{
A meta-analysis class object containing all results.
}
\description{
Taylor series approximation artifact-distribution meta-analysis correcting for Case V indirect range restriction and measurement error
}
\keyword{internal}
