% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_r_ad_rb_adj.R
\name{ma_r_ad.tsa_rb1Adj}
\alias{ma_r_ad.tsa_rb1Adj}
\title{Taylor series approximation artifact-distribution meta-analysis correcting for Raju and Burke's case 1 direct range restriction and measurement error}
\usage{
ma_r_ad.tsa_rb1Adj(x)
}
\arguments{
\item{x}{List of bare-bones meta-analytic data, artifact-distribution objects for X and Y, and other meta-analysis options.}
}
\value{
A list of artifact-distribution meta-analysis results to be returned to the ma_r_ad function.
}
\description{
Taylor series approximation artifact-distribution meta-analysis correcting for Raju and Burke's case 1 direct range restriction and measurement error
}
\references{
Raju, N. S., & Burke, M. J. (1983).
Two new procedures for studying validity generalization.
\emph{Journal of Applied Psychology, 68}(3), 382–395. \url{https://doi.org/10.1037/0021-9010.68.3.382}
}
\keyword{internal}
