% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_r_ad_uvdrr.R
\name{ma_r_ad.tsa_uvdrr}
\alias{ma_r_ad.tsa_uvdrr}
\title{Taylor series approximation artifact-distribution meta-analysis correcting for Case II direct range restriction and measurement error}
\usage{
ma_r_ad.tsa_uvdrr(x)
}
\arguments{
\item{x}{List of bare-bones meta-analytic data, artifact-distribution objects for X and Y, and other meta-analysis options.}
}
\value{
A meta-analysis class object containing all results.
}
\description{
Implements a newly derived TSA model based on the Case II formula for direct range restriction to estimate sd_rho.
}
\keyword{internal}
