% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_r_ad_uvirr.R
\name{ma_r_ad.tsa_uvirr}
\alias{ma_r_ad.tsa_uvirr}
\title{Taylor series approximation artifact-distribution meta-analysis correcting for Case II direct range restriction and measurement error}
\usage{
ma_r_ad.tsa_uvirr(x)
}
\arguments{
\item{x}{List of bare-bones meta-analytic data, artifact-distribution objects for X and Y, and other meta-analysis options.}
}
\value{
A meta-analysis class object containing all results.
}
\description{
Taylor series approximation artifact-distribution meta-analysis correcting for Case II direct range restriction and measurement error
}
\references{
Hunter, J. E., Schmidt, F. L., & Le, H. (2006).
Implications of direct and indirect range restriction for meta-analysis methods and findings.
\emph{Journal of Applied Psychology, 91}(3), 594–612. \url{https://doi.org/10.1037/0021-9010.91.3.594}
}
\keyword{internal}
