% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{manage_arglength}
\alias{manage_arglength}
\title{Check the length of x against the length of y and replicate z if necessary}
\usage{
manage_arglength(x, y)
}
\arguments{
\item{x}{Argument to be checked.}

\item{y}{Argument against which \code{x} should be checked.}
}
\value{
Error message or vector of values
}
\description{
Check the length of x against the length of y and replicate z if necessary
}
\keyword{internal}
