% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_d.R
\name{merge_simdat_d}
\alias{merge_simdat_d}
\title{Merge multiple "simdat_d" class objects}
\usage{
merge_simdat_d(...)
}
\arguments{
\item{...}{Collection of objects created by the "simulate_d_database" function. Simply enter the database objects as \code{merge_simdat_d}(data_obj1, data_obj2, data_obj_3).}
}
\value{
A merged database of class \code{simdat_d}
}
\description{
This function allows for multiple simulated databases from \code{\link{simulate_d_database}} to be merged together into a single database. Merged databases will be assigned moderator variable codes.
}
