% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mixture.R
\name{mix_dist}
\alias{mix_dist}
\title{Descriptive statistics for a mixture distribution}
\usage{
mix_dist(mean_vec, var_vec, n_vec, unbiased = TRUE, na.rm = FALSE)
}
\arguments{
\item{mean_vec}{Vector of sample means.}

\item{var_vec}{Vector of sample variances.}

\item{n_vec}{Vector of sample sizes.}

\item{unbiased}{Logical scalar determining whether variance should be unbiased (TRUE; default) or maximum-likelihood (FALSE).}

\item{na.rm}{Logical scalar determining whether to remove missing values prior to computing output (TRUE) or not (FALSE; default)}
}
\value{
The mean, pooled sample (within-sample) variance, variance of sample means (between-groups), and mixture (total sample) variance of the mixture sample data.
}
\description{
\loadmathjax
Compute descriptive statistics for a mixture distribution. This function returns the grand mean, the pooled sample variance (mean square within), variance of sample means (mean square between), portions of the total variance that are within and between groups, and mixture (total sample) variance of the mixture sample data.
}
\details{
The grand mean of a mixture distribution is computed as:

\mjdeqn{\mu=\frac{\Sigma_{i=1}^{k}\bar{x}_{i}n_{i}}{\Sigma_{i=1}^{k}n_{i}}}{mu = sum(mean_vec * n_vec) / sum(n_vec)}

where \mjeqn{\mu}{mu} is the grand mean, \mjeqn{\bar{x}_{i}}{mean_vec} represents the sample means, and \mjeqn{n_{i}}{n_vec} represents the sample sizes.


Maximum-likelihood mixture variances are computed as:
\mjdeqn{var_{pooled_{ML}}=MSW_{ML}=\frac{\Sigma_{i=1}^{k}\left(\bar{x}_{i}-\mu\right)n_{i}}{\Sigma_{i=1}^{k}n_{i}}}{var_pooled_ml = MSW_ml = sum(var_vec * n_vec) / sum(n_vec)}
\mjdeqn{var_{means_{ML}}=MSB_{ML}=\frac{\Sigma_{i=1}^{k}\left(\bar{x}_{i}-\mu\right)n_{i}}{k}}{var_means_ml = MSB_ml = sum(var_vec * n_vec) / sum(n_vec)}
\mjdeqn{var_{BG_{ML}}=\frac{\Sigma_{i=1}^{k}\left(\bar{x}_{i}-\mu\right)n_{i}}{\Sigma_{i=1}^{k}n_{i}}}{var_bg_ml = sum((mean_vec - mean_mixture)^2 * n_vec) / sum(n_vec)}
\mjdeqn{var_{WG_{ML}}=\frac{\Sigma_{i=1}^{k}v_{i}n_{i}}{\Sigma_{i=1}^{k}n_{i}}}{var_wg_ml = sum(var_vec * n_vec) / sum(n_vec)}
\mjdeqn{var_{mix_{ML}}=var_{BG_{ML}}+var_{WG_{ML}}}{var_mix_ml = var_bg_ml + var_wg_ml}

where \mjeqn{v_{i}}{var_vec} represents the sample variances.

Unbiased mixture variances are computed as:
\mjdeqn{var_{pooled_{Unbiased}}=MSW_{Unbiased}=\frac{\Sigma_{i=1}^{k}v_{i}\left(n_{i}-1\right)}{\left(\Sigma_{i=1}^{k}n_{i}\right)-k}}{var_pooled_unbiased = MSW_unbiased = sum(var_vec * (n_vec - 1)) / (sum(n_vec) - k)}
\mjdeqn{var_{means_{Unbiased}}=MSB_{Unbiased}=\frac{\Sigma_{i=1}^{k}\left(\bar{x}_{i}-\mu\right)n_{i}}{k-1}}{var_means_unbiased = MSB_unbiased = sum((mean_vec - mean_mixture)^2 * n_vec) / (k - 1)}
\mjdeqn{var_{BG_{Unbiased}}=\frac{\Sigma_{i=1}^{k}\left(\bar{x}_{i}-\mu\right)n_{i}}{\left(\Sigma_{i=1}^{k}n_{i}\right)-1}}{var_bg_unbiased = sum((mean_vec - mean_mixture)^2 * n_vec) / (sum(n_vec) - 1)}
\mjdeqn{var_{WG_{Unbiased}}=\frac{\Sigma_{i=1}^{k}v_{i}\left(n_{i}-1\right)}{\left(\Sigma_{i=1}^{k}n_{i}\right)-1}}{var_wg_unbiased = sum(var_vec * (n_vec - 1)) / (sum(n_vec) - 1)}
\mjdeqn{var_{mix_{Unbiased}}=var_{BG_{Unbiased}}+var_{WG_{Unbiased}}}{var_mix_unbiased = var_bg_unbiased + var_wg_unbiased}
}
\examples{
mix_dist(mean_vec = c(-.5, 0, .5), var_vec = c(.9, 1, 1.1), n_vec = c(100, 100, 100))
}
\keyword{univar}
