% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_predict.R
\name{predict.psychmeta}
\alias{predict.psychmeta}
\alias{predict.lm_mat}
\alias{predict.summary.lm_mat}
\title{Prediction method for objects of classes deriving from "lm_mat"}
\usage{
predict.psychmeta(...)

\method{predict}{lm_mat}(object, newdata, se.fit = FALSE, df = Inf,
  interval = "none", level = 0.95, ...)

\method{predict}{summary.lm_mat}(object, newdata, se.fit = FALSE, df = Inf,
  interval = "none", level = 0.95, ...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}

\item{object}{Object of class inheriting from "lm_mat"}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{se.fit}{A switch indicating if standard errors are required.}

\item{df}{Degrees of freedom for scale.}

\item{interval}{Type of interval calculation. Can be abbreviated.}

\item{level}{Tolerance/confidence level.}
}
\value{
An set of predicted values
}
\description{
Prediction method for objects of classes deriving from "lm_mat"
}
\keyword{internal}
