% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.ad_int}
\alias{print.ad_int}
\title{print method for interactive artifact distributions}
\usage{
\method{print}{ad_int}(x, ..., digits = 3)
}
\description{
print method for interactive artifact distributions
}
\keyword{internal}
