% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.get_escalc}
\alias{print.get_escalc}
\title{print method for lists of escalc objects retrieved with \code{get_escalc()}}
\usage{
\method{print}{get_escalc}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
print method for lists of escalc objects retrieved with \code{get_escalc()}
}
\keyword{internal}
