% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.get_leave1out}
\alias{print.get_leave1out}
\title{print method for lists of leave-one-out meta-analyses retrieved with \code{get_leave1out()}}
\usage{
\method{print}{get_leave1out}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
print method for lists of leave-one-out meta-analyses retrieved with \code{get_leave1out()}
}
\keyword{internal}
