% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.get_matrix}
\alias{print.get_matrix}
\title{print method for meta-analytic matrices retrieved with \code{get_matrix()}}
\usage{
\method{print}{get_matrix}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
print method for meta-analytic matrices retrieved with \code{get_matrix()}
}
\keyword{internal}
