% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.get_plots}
\alias{print.get_plots}
\title{print method for meta-analysis plots retrieved with \code{get_plots()}}
\usage{
\method{print}{get_plots}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
print method for meta-analysis plots retrieved with \code{get_plots()}
}
\keyword{internal}
