% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matreg.R
\name{print.lm_mat}
\alias{print.lm_mat}
\title{Print method for objects of the class "lm_mat "}
\usage{
\method{print}{lm_mat}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\description{
Print method for objects of the class "lm_mat "
}
\keyword{internal}
