% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.psychmeta.correct_r}
\alias{print.psychmeta.correct_r}
\title{print method for correlations corrected for artifacts}
\usage{
\method{print}{psychmeta.correct_r}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
print method for correlations corrected for artifacts
}
\keyword{internal}
