% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.psychmeta.get_bootstrap}
\alias{print.psychmeta.get_bootstrap}
\title{print method for lists of bootstrapped meta-analyses retrieved with \code{get_bootstrap()}}
\usage{
\method{print}{psychmeta.get_bootstrap}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
print method for lists of bootstrapped meta-analyses retrieved with \code{get_bootstrap()}
}
\keyword{internal}
