% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.psychmeta.ma_d.ad}
\alias{print.psychmeta.ma_d.ad}
\title{print method for artifact-distribution meta-analyses of d values}
\usage{
\method{print}{psychmeta.ma_d.ad}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
print method for artifact-distribution meta-analyses of d values
}
\keyword{internal}
