% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.psychmeta.ma_d.order2}
\alias{print.psychmeta.ma_d.order2}
\title{print method for second-order meta-analyses of d values}
\usage{
\method{print}{psychmeta.ma_d.order2}(x, ..., digits = 3, verbose = FALSE)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}

\item{verbose}{Logical scalar determining whether objects with verbose-printing options should be printed in great detail (\code{TRUE}) or succinctly (\code{FALSE}; default).}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
print method for second-order meta-analyses of d values
}
\keyword{internal}
