% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.psychmeta.ma_r.ic.master}
\alias{print.psychmeta.ma_r.ic.master}
\title{print method for individual-correction meta-analyses of correlations}
\usage{
\method{print}{psychmeta.ma_r.ic.master}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
print method for individual-correction meta-analyses of correlations
}
\keyword{internal}
