% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.psychmeta.simulate_psych}
\alias{print.psychmeta.simulate_psych}
\title{print method for simulated psychometric studies}
\usage{
\method{print}{psychmeta.simulate_psych}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
print method for simulated psychometric studies
}
\keyword{internal}
