% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mat.R
\name{print.summary.lm_mat}
\alias{print.summary.lm_mat}
\title{Print method for objects of the class "summary.lm_mat"}
\usage{
\method{print}{summary.lm_mat}(x, digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{an object of class "summary.lm_mat", usually, a result of a call to summary.lm_mat.}

\item{digits}{Number of digits to which result should be rounded}

\item{symbolic.cor}{logical. If TRUE, print the correlations in a symbolic form (see symnum) rather than as numbers.}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{...}{Additional arguments}
}
\value{
Regression results printed in console
}
\description{
Print method for objects of the class "summary.lm_mat"
}
