% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matreg.R
\name{print.summary_lm_mat}
\alias{print.summary_lm_mat}
\title{Print method for objects of the class "summary_lm_mat "}
\usage{
\method{print}{summary_lm_mat}(x, digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"), ...)
}
\description{
Print method for objects of the class "summary_lm_mat "
}
\keyword{internal}
