% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{psychmeta_news}
\alias{psychmeta_news}
\title{Retrieve the \code{NEWS} file for the \emph{psychmeta} package}
\usage{
psychmeta_news()
}
\description{
Displays \emph{psychmeta}'s \code{NEWS} file, which contains version information, outlines additions and changes to the package, and describes other updates.
}
\examples{
psychmeta_news()
}
