% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_reattribute.R
\name{reattribute}
\alias{reattribute}
\title{Copy class and attributes from the original version of an object to a modified version.}
\usage{
reattribute(x, result)
}
\arguments{
\item{x}{The original object, which has a class/attributes to copy}

\item{result}{The modified object, which is / might be missing the class/attributes.}
}
\value{
\code{result}, now with class/attributes restored.
}
\description{
Copy class and attributes from the original version of an object to a modified version.
}
