% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{record_warnings}
\alias{record_warnings}
\title{Summary of warning messages generated within a function}
\usage{
record_warnings()
}
\value{
A data frame containing a summary of warning messages and their frequencies.
}
\description{
Summary of warning messages generated within a function
}
\keyword{internal}
