% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{round2char}
\alias{round2char}
\title{Round numeric values to an exact number of digits and return as a character}
\usage{
round2char(x, digits = 3, na_replace = "", omit_leading_zero = FALSE)
}
\arguments{
\item{x}{Numeric values}

\item{digits}{Number of digits to which result should be rounded}

\item{na_replace}{Scalar value: Character with which NA values should be replaced}

\item{omit_leading_zero}{Logical scalar determining whether to omit leading zeros (\code{TRUE}) or retain them (\code{FALSE}; default).}
}
\value{
A vector of rounded numbers converted to characters
}
\description{
Round numeric values to an exact number of digits and return as a character
}
\examples{
# round2char(x = .50000005)
# round2char(x = NA, na_replace = "---")
}
\keyword{internal}
