% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{scalar_arg_warning}
\alias{scalar_arg_warning}
\title{Warning message for scalar arguments receiving multiple values}
\usage{
scalar_arg_warning(arg, arg_name = NULL)
}
\arguments{
\item{arg}{Argument value}

\item{arg_name}{Argument name}
}
\value{
Warning if length of arg is greater than 1 and the first element of arg.
}
\description{
Warning message for scalar arguments receiving multiple values
}
\keyword{internal}
