% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{screen_rel}
\alias{screen_rel}
\title{Screen to detect impossible values in vectors of reliability estimates.}
\usage{
screen_rel(rel_vec, art_name = "Reliability")
}
\arguments{
\item{rel_vec}{Vector of reliability estimates.}

\item{art_name}{Optional artifact name to use in warning messages.}
}
\value{
Does not return values; stops processes if improper values are used.
}
\description{
Screen to detect impossible values in vectors of reliability estimates.
}
\examples{
## Not run
## screen_rel(rel_vec = c(.8, Inf), art_name = "rxxa")
## screen_rel(rel_vec = c(.8, -.2), art_name = "rxxa")
}
\keyword{internal}
