% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{screen_u}
\alias{screen_u}
\title{Screen to detect impossible values in vectors of u ratios.}
\usage{
screen_u(u_vec, art_name = "u ratio")
}
\arguments{
\item{u_vec}{Vector of u ratios.}

\item{art_name}{Optional artifact name to use in warning messages.}
}
\value{
Does not return values; stops processes if improper values are used.
}
\description{
Screen to detect impossible values in vectors of u ratios.
}
\examples{
## Not run
## screen_u(u_vec = c(0, .8), art_name = "ux")
## screen_u(u_vec = c(-1, .8), art_name = "ux")
## screen_u(u_vec = c(Inf, .8), art_name = "ux")
}
\keyword{internal}
