% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mat.R
\name{summary.lm_mat}
\alias{summary.lm_mat}
\title{Summary method for objects of the class "lm_mat"}
\usage{
\method{summary}{lm_mat}(object, ...)
}
\arguments{
\item{object}{Regression object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Summary of regression model
}
\description{
Summary method for objects of the class "lm_mat"
}
