% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summary.R
\name{summary.psychmeta}
\alias{summary.psychmeta}
\alias{summary.ma_psychmeta}
\alias{summary.lm_mat}
\alias{summary.ad_tsa}
\alias{summary.ad_int_list}
\alias{summary.ad_int}
\alias{summary.correct_r}
\alias{summary.correct_d}
\alias{summary.simdat_psych}
\alias{summary.simdat_r_sample}
\alias{summary.simdat_r_database}
\alias{summary.simdat_d_sample}
\alias{summary.simdat_d_database}
\alias{summary.convert_es}
\alias{summary.dmod}
\alias{summary.ma_heterogeneity}
\alias{summary.ma_leave1out}
\alias{summary.ma_cumulative}
\alias{summary.ma_bootstrap}
\alias{summary.get_metatab}
\alias{summary.get_plots}
\alias{summary.get_matrix}
\alias{summary.get_escalc}
\alias{summary.get_followup}
\alias{summary.get_heterogeneity}
\alias{summary.get_metareg}
\alias{summary.get_bootstrap}
\alias{summary.get_leave1out}
\alias{summary.get_cumulative}
\alias{summary.ad_list}
\alias{summary.ma_table}
\alias{summary.ma_ic_list}
\alias{summary.ma_ad_list}
\title{Summary methods for \pkg{psychmeta}}
\usage{
\method{summary}{ma_psychmeta}(object, ...)

\method{summary}{lm_mat}(object, ...)

\method{summary}{ad_tsa}(object, ...)

\method{summary}{ad_int_list}(object, ...)

\method{summary}{ad_int}(object, ...)

\method{summary}{correct_r}(object, ...)

\method{summary}{correct_d}(object, ...)

\method{summary}{simdat_psych}(object, ...)

\method{summary}{simdat_r_sample}(object, ...)

\method{summary}{simdat_r_database}(object, ...)

\method{summary}{simdat_d_sample}(object, ...)

\method{summary}{simdat_d_database}(object, ...)

\method{summary}{convert_es}(object, ...)

\method{summary}{dmod}(object, ...)

\method{summary}{ma_heterogeneity}(object, ...)

\method{summary}{ma_leave1out}(object, ...)

\method{summary}{ma_cumulative}(object, ...)

\method{summary}{ma_bootstrap}(object, ...)

\method{summary}{get_metatab}(object, ...)

\method{summary}{get_plots}(object, ...)

\method{summary}{get_matrix}(object, ...)

\method{summary}{get_escalc}(object, ...)

\method{summary}{get_followup}(object, ...)

\method{summary}{get_heterogeneity}(object, ...)

\method{summary}{get_metareg}(object, ...)

\method{summary}{get_bootstrap}(object, ...)

\method{summary}{get_leave1out}(object, ...)

\method{summary}{get_cumulative}(object, ...)

\method{summary}{ad_list}(object, ...)

\method{summary}{ma_table}(object, ...)

\method{summary}{ma_ic_list}(object, ...)

\method{summary}{ma_ad_list}(object, ...)
}
\arguments{
\item{object}{Object to be printed (object is used to select a method).}

\item{...}{Additional arguments.}
}
\value{
Summary object.
}
\description{
Summary methods for \pkg{psychmeta} output objects with classes exported from \pkg{psychmeta}.
}
\keyword{internal}
