% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summary.R
\name{summary}
\alias{summary}
\title{Summary methods for \pkg{psychmeta}}
\arguments{
\item{object}{Object to be printed (object is used to select a method).}

\item{...}{Additional arguments.}
}
\value{
Summary object.
}
\description{
Summary methods for \pkg{psychmeta} output objects with classes exported from \pkg{psychmeta}.
}
