% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{warning_variance}
\alias{warning_variance}
\title{Warning message for invalid variances}
\usage{
warning_variance(var, var_name = NULL, sd_warning = TRUE)
}
\arguments{
\item{var}{Variance object}

\item{var_name}{Name of variance object}
}
\value{
A warning, if the supplied variance does not produce a valid standard deviation
}
\description{
Warning message for invalid variances
}
\keyword{internal}
