% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mat.R
\name{anova.lm_mat}
\alias{anova.lm_mat}
\title{Anova method for objects of the class "lm_mat"}
\usage{
\method{anova}{lm_mat}(...)
}
\arguments{
\item{...}{Arguments}
}
\value{
An anova table
}
\description{
Anova method for objects of the class "lm_mat"
}
