% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_confint.R
\name{confint}
\alias{confint}
\alias{confint.lm_mat}
\alias{confint.summary.lm_mat}
\title{Confidence interval method for objects of classes deriving from "lm_mat"}
\usage{
\method{confint}{lm_mat}(object, parm, level = 0.95, ...)

\method{confint}{summary.lm_mat}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{Matrix regression object.}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{Confidence level}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Lower and upper bounds of confidence intervals for regression coefficients.
}
\description{
Confidence interval method for objects of classes deriving from "lm_mat"
}
\keyword{internal}
