% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_r_gonzalezmule_2014}
\alias{data_r_gonzalezmule_2014}
\title{Meta-analysis of OCB correlations with other constructs}
\format{
data.frame
}
\usage{
data(data_r_gonzalezmule_2014)
}
\description{
Data set to demonstrate corrections for univariate range restriction and measurement error using
individual corrections or artifact distributions.
NOTE: This is an updated version of the data set reported in the Gonzalez-Mulé, Mount, an Oh (2014) article that was obtained from the first author.
}
\examples{
data(data_r_gonzalezmule_2014)
}
\references{
Gonzalez-Mulé, E., Mount, M. K., & Oh, I.-S. (2014).
A meta-analysis of the relationship between general mental ability and nontask performance.
\emph{Journal of Applied Psychology, 99}(6), 1222–1243. \url{https://doi.org/10.1037/a0037547}
}
\keyword{datasets}
