% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.plot_funnel}
\alias{.plot_funnel}
\title{Internal funnel-plot generator}
\usage{
.plot_funnel(x, label_es = "Effect Size", conf_level = c(0.95, 0.99),
  conf_linetype = c("dashed", "dotted"), conf_fill = NA,
  conf_alpha = 1, null_effect = NA, null_conf_level = c(0.9, 0.95,
  0.99), null_conf_linetype = c("solid", "dashed", "dotted"),
  null_conf_fill = "black", null_conf_alpha = c(0, 0.2, 0.4))
}
\arguments{
\item{x}{An escalc-class object}
}
\value{
A funnel plot.
}
\description{
Internal funnel-plot generator
}
\author{
John Sakaluk and Brenton Wiernik
}
\keyword{internal}
