% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_r.R
\name{format_long}
\alias{format_long}
\title{Create long-format datasets in simulate_r_database}
\usage{
format_long(x, param, sample_id, var_names, show_applicant, decimals = 2,
  noalpha = FALSE)
}
\arguments{
\item{x}{Simulation list}

\item{param}{Is simulation data parameter data (TRUE) or sample data (FALSE)?}

\item{sample_id}{What is the ID associated with the sample?}

\item{var_names}{Variables to pull from simulation list}

\item{show_applicant}{Should applicant data be shown for sample statistics (TRUE) or suppressed (FALSE)?}

\item{decimals}{Number of decimals to which statistical results (not parameters) should be rounded. Rounding to 2 decimal places best captures the precision of data available from published primary research.}

\item{noalpha}{Logical scalar indicating whether or not alpha is in the database.}
}
\value{
A dataframe of results
}
\description{
Create long-format datasets in simulate_r_database
}
\keyword{internal}
