% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_d_order2.R
\name{ma_d_order2}
\alias{ma_d_order2}
\title{Second-order meta-analysis function for \emph{d} values}
\usage{
ma_d_order2(k, N = NULL, d = NULL, delta = NULL, var_d = NULL,
  var_d_c = NULL, ma_type = c("bb", "ic", "ad"), sample_id = NULL,
  citekey = NULL, moderators = NULL, moderator_type = "simple",
  construct_x = NULL, construct_y = NULL, data = NULL,
  control = control_psychmeta(), ...)
}
\arguments{
\item{k}{Vector or column name of meta-analyses' k values.}

\item{N}{Vector or column name of meta-analyses' total sample sizes (optional).}

\item{d}{Vector or column name of mean observed \emph{d} values.}

\item{delta}{Vector or column name of mean corrected \emph{d} values.}

\item{var_d}{Vector or column name of observed variances of observed \emph{d} values.}

\item{var_d_c}{Vector or column name of observed variances of corrected \emph{d} values.}

\item{ma_type}{Type of meta-analyses being analyzed: "bb" (barebones), "ic" (individual correction), or "ad" (artifact distribution).}

\item{sample_id}{Vector or column name of study ID labels.}

\item{citekey}{Optional vector of bibliographic citation keys for samples/studies in the meta-analysis (if multiple citekeys pertain to a given effect size, combine them into a single string entry with comma delimiters (e.g., "citkey1,citekey2").}

\item{moderators}{Matrix or column names of moderator variables to be used in the meta-analysis (can be a vector in the case of one moderator).}

\item{moderator_type}{Type of moderator analysis ("none", "simple", or "hierarchical").}

\item{construct_x}{Vector or column name of construct names for X.}

\item{construct_y}{Vector or column name of construct names for Y.}

\item{data}{Data frame containing columns whose names may be provided as arguments to vector arguments and/or moderators.}

\item{control}{Output from the \code{control_psychmeta()} function or a list of arguments controlled by the \code{control_psychmeta()} function. Ellipsis arguments will be screened for internal inclusion in \code{control}.}

\item{...}{Further arguments to be passed to functions called within the meta-analysis.}
}
\value{
A nested tabular object of the class "ma_psychmeta".
}
\description{
This function computes second-order meta-analysis function for \emph{d} values. It supports second-order analyses of bare-bones, artifact-distribution, and individual-correction meta-analyses.
}
