% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mat.R
\name{print.lm_mat}
\alias{print.lm_mat}
\title{Print method for objects of the class "lm_mat"}
\usage{
\method{print}{lm_mat}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Output from the \code{lm_mat()} function.}

\item{digits}{Number of digits to which result should be rounded}

\item{...}{Additional arguments}
}
\value{
Regression results printed in console
}
\description{
Print method for objects of the class "lm_mat"
}
