% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.psychmeta}
\alias{print.psychmeta}
\title{Master print method for objects of the 'psychmeta' class}
\usage{
\method{print}{psychmeta}(x, ..., digits = 3, verbose = FALSE)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}

\item{digits}{Number of digits to which results should be printed.}

\item{verbose}{Logical scalar determining whether objects with verbose-printing options should be printed in great detail (\code{TRUE}) or succinctly (\code{FALSE}; default).}
}
\value{
Printed results from objects of the 'psychmeta' class.
}
\description{
Master print method for objects of the 'psychmeta' class
}
\keyword{internal}
