% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_print.R
\name{print.psychmeta}
\alias{print.psychmeta}
\alias{print.lm_mat}
\alias{print.summary.lm_mat}
\alias{print.ad_tsa}
\alias{print.ad_int_list}
\alias{print.ad_int}
\alias{print.correct_r}
\alias{print.correct_d}
\alias{print.simdat_psych}
\alias{print.simdat_r_sample}
\alias{print.simdat_r_database}
\alias{print.simdat_d_sample}
\alias{print.simdat_d_database}
\alias{print.convert_es}
\alias{print.dmod}
\alias{print.ma_heterogeneity}
\alias{print.ma_leave1out}
\alias{print.ma_cumulative}
\alias{print.ma_bootstrap}
\alias{print.get_metatab}
\alias{print.get_plots}
\alias{print.get_matrix}
\alias{print.get_escalc}
\alias{print.get_followup}
\alias{print.get_heterogeneity}
\alias{print.get_metareg}
\alias{print.get_bootstrap}
\alias{print.get_leave1out}
\alias{print.get_cumulative}
\alias{print.ad_list}
\alias{print.ma_psychmeta}
\alias{print.ma_table}
\alias{print.ma_ic_list}
\alias{print.ma_ad_list}
\alias{print.summary.ma_psychmeta}
\title{Print methods for \pkg{psychmeta}}
\usage{
print.psychmeta(...)

\method{print}{lm_mat}(x, ..., digits = max(3L, getOption("digits") - 3L))

\method{print}{summary.lm_mat}(x, digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"), ...)

\method{print}{ad_tsa}(x, ..., digits = 3)

\method{print}{ad_int_list}(x, ..., digits = 3)

\method{print}{ad_int}(x, ..., digits = 3)

\method{print}{correct_r}(x, ..., digits = 3)

\method{print}{correct_d}(x, ..., digits = 3)

\method{print}{simdat_psych}(x, ..., digits = 3)

\method{print}{simdat_r_sample}(x, ..., digits = 3)

\method{print}{simdat_r_database}(x, ..., digits = 3)

\method{print}{simdat_d_sample}(x, ..., digits = 3)

\method{print}{simdat_d_database}(x, ..., digits = 3)

\method{print}{convert_es}(x, ..., digits = 3)

\method{print}{dmod}(x, ..., digits = 3)

\method{print}{ma_heterogeneity}(x, ..., digits = 3)

\method{print}{ma_leave1out}(x, ..., digits = 3)

\method{print}{ma_cumulative}(x, ..., digits = 3)

\method{print}{ma_bootstrap}(x, ..., digits = 3)

\method{print}{get_metatab}(x, ..., digits = 3)

\method{print}{get_plots}(x, ..., digits = 3)

\method{print}{get_matrix}(x, ..., digits = 3)

\method{print}{get_escalc}(x, ..., digits = 3)

\method{print}{get_followup}(x, ..., digits = 3)

\method{print}{get_heterogeneity}(x, ..., digits = 3)

\method{print}{get_metareg}(x, ..., digits = 3)

\method{print}{get_bootstrap}(x, ..., digits = 3)

\method{print}{get_leave1out}(x, ..., digits = 3)

\method{print}{get_cumulative}(x, ..., digits = 3)

\method{print}{ad_list}(x, ..., digits = 3)

\method{print}{ma_psychmeta}(x, ..., digits = 3, verbose = FALSE)

\method{print}{ma_table}(x, ..., digits = 3, verbose = FALSE)

\method{print}{ma_ic_list}(x, ..., digits = 3, verbose = FALSE)

\method{print}{ma_ad_list}(x, ..., digits = 3, verbose = FALSE)

\method{print}{summary.ma_psychmeta}(x, ..., ma_methods = NULL,
  correction_types = "ts", verbose = FALSE)
}
\arguments{
\item{...}{Additional arguments.}

\item{x}{Object to be printed (object is used to select a method).}

\item{digits}{Number of digits to which results should be rounded.}

\item{symbolic.cor}{For \code{lm_mat} output: logical. If TRUE, print the correlations in a symbolic form (see symnum) rather than as numbers.}

\item{signif.stars}{For \code{lm_mat} output: logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{verbose}{Logical scalar that determines whether printed object should contain verbose information (e.g., non-standard columns of meta-analytic output; \code{TRUE}) or not (\code{FALSE}).}

\item{ma_methods}{Meta-analytic methods to be included. Valid options are: "bb", "ic", and "ad"}

\item{correction_types}{Types of meta-analytic corrections to be incldued. Valid options are: "ts", "vgx", and "vgy"}
}
\value{
Printed object.
}
\description{
Print methods for \pkg{psychmeta} output objects with classes exported from \pkg{psychmeta}.
}
\keyword{internal}
