% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_matrix.R
\name{unmix_matrix}
\alias{unmix_matrix}
\title{Estimate average within-group covariance matrices from a mixture covariance matrix}
\usage{
unmix_matrix(sigma, mu_mat, p_vec, N = NULL, group_names = NULL,
  var_names = NULL)
}
\arguments{
\item{sigma}{Mixture covariance matrix.}

\item{mu_mat}{Matrix of mean parameters, with groups on the rows and variables on the columns.}

\item{p_vec}{Vector of proportion of cases in each group.}

\item{N}{Optional total sample size across all groups (used to compute unbiased covariance estimates).}

\item{group_names}{Optional vector of group names.}

\item{var_names}{Optional vector of variable names.}
}
\value{
List of within-group covariances and means.
}
\description{
Estimate average within-group covariance matrices from a mixture covariance matrix
}
